Consolidate router components
