+++ src/components/App/component.tsx
@@ -1,9 +1,9 @@
 import { ReactElement, Suspense } from "react";
 import { RouterProvider } from "react-router-dom";
 
-import router from "browserRouter";
 import AnnouncementBanner from "components/AnnouncementBanner/AnnouncementBanner";
 import UpperRightToastContainer from "components/Toast/UpperRightToastContainer";
+import router from "router/browserRouter";
 
 /**
  * The top-level component
