+++ src/components/Login/tests/Signup.test.tsx
@@ -23,10 +23,10 @@ jest.mock(
 jest.mock("backend", () => ({
   getBannerText: () => Promise.resolve(""),
 }));
-jest.mock("browserRouter");
 jest.mock("components/Login/Redux/LoginActions", () => ({
   asyncSignUp: (...args: any[]) => mockAsyncSignUp(...args),
 }));
+jest.mock("router/browserRouter");
 jest.mock("types/hooks", () => {
   return {
     ...jest.requireActual("types/hooks"),
