+++ src/components/ProjectScreen/CreateProjectActions.ts
@@ -1,8 +1,8 @@
 import { WritingSystem } from "api/models";
 import { createProject, finishUploadLift, getProject } from "backend";
-import router from "browserRouter";
 import { asyncCreateUserEdits } from "components/GoalTimeline/Redux/GoalActions";
 import { setNewCurrentProject } from "components/Project/ProjectActions";
+import router from "router/browserRouter";
 import { StoreStateDispatch } from "types/Redux/actions";
 import { Path } from "types/path";
 import { newProject } from "types/project";
